# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Проводник 400",
    "description": "Тренажёр нештатных ситуаций для проводников ВСМ-400.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/integration/employees/{code}/results": {
      "get": {
        "tags": [
          "Интеграция"
        ],
        "summary": "Результаты сотрудника: уровень, владение, прохождения, достижения",
        "operationId": "employee_results_api_integration_employees__code__results_get",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeResultsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/integration/results": {
      "get": {
        "tags": [
          "Интеграция"
        ],
        "summary": "Экспорт завершённых прохождений с курсором",
        "operationId": "results_api_integration_results_get",
        "parameters": [
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 9223372036854775807,
              "minimum": 0,
              "default": 0,
              "title": "Cursor"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultsPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/integration/events": {
      "get": {
        "tags": [
          "Интеграция"
        ],
        "summary": "События для LMS с курсором after_id",
        "operationId": "events_api_integration_events_get",
        "parameters": [
          {
            "name": "after_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 9223372036854775807,
              "minimum": 0,
              "default": 0,
              "title": "After Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventsPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/integration/events/ack": {
      "post": {
        "tags": [
          "Интеграция"
        ],
        "summary": "Подтвердить доставку событий",
        "operationId": "ack_api_integration_events_ack_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AckRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AckResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/integration/competencies": {
      "get": {
        "tags": [
          "Интеграция"
        ],
        "summary": "Справочник компетенций",
        "operationId": "competencies_api_integration_competencies_get",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CompetencyRef"
                  },
                  "title": "Response Competencies Api Integration Competencies Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/integration/employees/{code}/actions": {
      "get": {
        "tags": [
          "Интеграция"
        ],
        "summary": "Журнал действий сотрудника с курсором after_id",
        "operationId": "actions_api_integration_employees__code__actions_get",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          },
          {
            "name": "after_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 9223372036854775807,
              "minimum": 0,
              "default": 0,
              "title": "After Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionsPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/integration/employees": {
      "post": {
        "tags": [
          "Интеграция"
        ],
        "summary": "Создать сотрудника из HR-системы; PIN возвращается один раз",
        "operationId": "create_employee_api_integration_employees_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEmployeeRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedEmployeeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/scenarios/reload": {
      "post": {
        "tags": [
          "Администрирование"
        ],
        "summary": "Перечитать сценарии и справочники, уведомить о новых",
        "operationId": "reload_api_admin_scenarios_reload_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReloadReport"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/admin/scenarios/validate": {
      "get": {
        "tags": [
          "Администрирование"
        ],
        "summary": "Проверить файлы контента валидатором без применения",
        "operationId": "validate_api_admin_scenarios_validate_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReloadReport"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "AchievementRef": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "earned_at": {
            "type": "string",
            "title": "Earned At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "earned_at"
        ],
        "title": "AchievementRef"
      },
      "AckRequest": {
        "properties": {
          "ids": {
            "items": {
              "type": "integer",
              "maximum": 9.223372036854776e+18,
              "minimum": 1.0
            },
            "type": "array",
            "maxItems": 500,
            "minItems": 1,
            "title": "Ids"
          }
        },
        "type": "object",
        "required": [
          "ids"
        ],
        "title": "AckRequest"
      },
      "AckResponse": {
        "properties": {
          "acked": {
            "type": "integer",
            "title": "Acked"
          }
        },
        "type": "object",
        "required": [
          "acked"
        ],
        "title": "AckResponse"
      },
...
      "ActionsPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ActionOut"
            },
            "type": "array",
            "title": "Items"
          },
          "next_after_id": {
            "type": "integer",
            "title": "Next After Id"
          }
        },
        "type": "object",
        "required": [
          "items",
          "next_after_id"
        ],
        "title": "ActionsPage"
      },
...
      "CompetencyRef": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "code",
          "title",
          "description"
        ],
        "title": "CompetencyRef"
      },
...
      "CreateEmployeeRequest": {
        "properties": {
          "employee_code": {
            "type": "string",
            "maxLength": 32,
            "minLength": 3,
            "pattern": "^[A-Za-z0-9-]+$",
            "title": "Employee Code"
          },
          "display_name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Display Name"
          },
          "brigade": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Brigade"
          },
          "role": {
            "type": "string",
            "enum": [
              "conductor",
              "mentor"
            ],
            "title": "Role",
            "default": "conductor"
          }
        },
        "type": "object",
        "required": [
          "employee_code",
          "display_name",
          "brigade"
        ],
        "title": "CreateEmployeeRequest"
      },
      "CreatedEmployeeOut": {
        "properties": {
          "employee_code": {
            "type": "string",
            "title": "Employee Code"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "brigade": {
            "type": "string",
            "title": "Brigade"
          },
          "depot": {
            "type": "string",
            "title": "Depot"
          },
          "pin": {
            "type": "string",
            "title": "Pin"
          }
        },
        "type": "object",
        "required": [
          "employee_code",
          "display_name",
          "role",
          "brigade",
          "depot",
          "pin"
        ],
        "title": "CreatedEmployeeOut"
      },
...
      "EmployeeResultsOut": {
        "properties": {
          "employee_code": {
            "type": "string",
            "title": "Employee Code"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "brigade": {
            "type": "string",
            "title": "Brigade"
          },
          "depot": {
            "type": "string",
            "title": "Depot"
          },
          "xp_total": {
            "type": "integer",
            "title": "Xp Total"
          },
          "level": {
            "$ref": "#/components/schemas/LevelOut"
          },
          "competencies": {
            "items": {
              "$ref": "#/components/schemas/IntegrationCompetencyOut"
            },
            "type": "array",
            "title": "Competencies"
          },
          "runs": {
            "items": {
              "$ref": "#/components/schemas/ResultItem"
            },
            "type": "array",
            "title": "Runs"
          },
          "achievements": {
            "items": {
              "$ref": "#/components/schemas/AchievementRef"
            },
            "type": "array",
            "title": "Achievements"
          }
        },
        "type": "object",
        "required": [
          "employee_code",
          "display_name",
          "role",
          "brigade",
          "depot",
          "xp_total",
          "level",
          "competencies",
          "runs",
          "achievements"
        ],
        "title": "EmployeeResultsOut"
      },
...
      "EventsPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EventOut"
            },
            "type": "array",
            "title": "Items"
          },
          "next_after_id": {
            "type": "integer",
            "title": "Next After Id"
          }
        },
        "type": "object",
        "required": [
          "items",
          "next_after_id"
        ],
        "title": "EventsPage"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IntegrationCompetencyOut": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "mastery": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mastery"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "runs_assessed": {
            "type": "integer",
            "title": "Runs Assessed"
          }
        },
        "type": "object",
        "required": [
          "code",
          "mastery",
          "status",
          "runs_assessed"
        ],
        "title": "IntegrationCompetencyOut"
      },
...
      "LevelOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "threshold": {
            "type": "integer",
            "title": "Threshold"
          },
          "next_threshold": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Threshold"
          },
          "next_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Title"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "threshold",
          "next_threshold",
          "next_title"
        ],
        "title": "LevelOut"
      },
...
      "ReloadReport": {
        "properties": {
          "loaded": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Loaded"
          },
          "new": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "New"
          },
          "removed": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Removed"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ContentErrorOut"
            },
            "type": "array",
            "title": "Errors"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "notified": {
            "type": "integer",
            "title": "Notified"
          }
        },
        "type": "object",
        "required": [
          "loaded",
          "new",
          "removed",
          "errors",
          "summary",
          "notified"
        ],
        "title": "ReloadReport"
      },
      "ResultItem": {
        "properties": {
          "run_id": {
            "type": "integer",
            "title": "Run Id"
          },
          "employee_code": {
            "type": "string",
            "title": "Employee Code"
          },
          "brigade": {
            "type": "string",
            "title": "Brigade"
          },
          "scenario_id": {
            "type": "string",
            "title": "Scenario Id"
          },
          "scenario_version": {
            "type": "integer",
            "title": "Scenario Version"
          },
          "finished_at": {
            "type": "string",
            "title": "Finished At"
          },
          "outcome": {
            "type": "string",
            "title": "Outcome"
          },
          "loyalty_final": {
            "type": "integer",
            "title": "Loyalty Final"
          },
          "safety_final": {
            "type": "integer",
            "title": "Safety Final"
          },
          "score": {
            "type": "integer",
            "title": "Score"
          },
          "xp": {
            "type": "integer",
            "title": "Xp"
          },
          "competencies": {
            "additionalProperties": true,
            "type": "object",
            "title": "Competencies"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "employee_code",
          "brigade",
          "scenario_id",
          "scenario_version",
          "finished_at",
          "outcome",
          "loyalty_final",
          "safety_final",
          "score",
          "xp",
          "competencies"
        ],
        "title": "ResultItem"
      },
      "ResultsPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ResultItem"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "ResultsPage"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}